requirements.txt
click==8.1.3
clip==1.0
keras==2.12.0rc1
matplotlib==3.7.1
numpy==1.23.5
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.4.0
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.28.2
scikit_learn==1.2.2
seaborn==0.12.2
setuptools==67.6.0
tensorflow==2.12.0
tensorflow_intel==2.12.0rc1
torch==2.0.0
pre-commit==3.2.1
git+https://github.com/openai/CLIP.git#egg=clip
